+++ src/fhirpath/fhirspec/__init__.py
@@ -10,7 +10,11 @@ from fhirspec import Configuration, FHIRStructureDefinition
 from fhirpath.enums import FHIR_VERSION
 from fhirpath.storage import FHIR_RESOURCE_SPEC_STORAGE, SEARCH_PARAMETERS_STORAGE
 
-from .spec import FHIRSearchSpec  # noqa: F401
+from .spec import (  # noqa: F401
+    FHIRSearchSpec,
+    SearchParameter,
+    ResourceSearchParameterDefinition,
+)
 from .spec import logger
 
 __author__ = "Md Nazrul Islam<email2nazrul@gmail.com>"